refactor: rename startup to http
