+++ crates/saldo-ws/src/lib.rs
@@ -1,3 +1,3 @@
+pub mod http;
 pub mod routes;
-pub mod startup;
 pub use sblex_telemetry as telemetry;
